Refactor: PipelineManager
