+++ app/globals.h
@@ -10,3 +10,20 @@ extern int dockedSizeX;
 extern float simSpeed;
 extern unsigned int collissionCount;
 extern float stiffness;
+
+struct WindowSize{
+    int width;
+    int height;
+    bool operator==(const WindowSize& other) const {
+        if(other.height == height && other.width == width){
+            return true;
+        }
+        return false;
+    }
+    bool operator!=(const WindowSize& other) const {
+        if(other.height == height && other.width == width){
+            return false;
+        }
+        return true;
+    }
+};
