+++ include/ThING/graphics/pipelineManager.h
@@ -1,7 +1,9 @@
 #pragma once
 
+#include <array>
 #include <filesystem>
 #include <fstream>
+#include <span>
 #include <string>
 #include <ThING/extras/handMade.h>
 #include <ThING/consts.h>
@@ -17,7 +19,6 @@
 class PipelineManager{
 public:
     PipelineManager();
-    PipelineManager(VkDevice device, VkFormat &swapChainImageFormat);
     ~PipelineManager();
     void init(VkDevice device, VkFormat &format);
     void createPipelines();
@@ -26,10 +27,10 @@ public:
 
     void cleanUp();
 
-    inline VkRenderPass& getrenderPass() {return renderPass;};
-    inline VkPipelineLayout* getLayouts() {return pipelineLayouts;};
-    inline VkPipeline* getGraphicsPipelines() {return graphicsPipelines;};
-    inline VkDescriptorSet* getDescriptorSets() {return descriptorSets;};
+    inline VkRenderPass& getRenderPass() {return renderPass;};
+    inline std::span<const VkPipelineLayout> getLayouts() const {return pipelineLayouts;};
+    inline std::span<const VkPipeline> getGraphicsPipelines() const {return graphicsPipelines;};
+    inline std::span<const VkDescriptorSet> getDescriptorSets() const {return descriptorSets;};
 private:
     void createDescriptorSetLayout();
     void createBasicGraphicsPipeline();
@@ -41,14 +42,14 @@ private:
     VkShaderModule createShaderModule(const std::vector<char>& code);
 
     VkRenderPass renderPass;
-    VkPipelineLayout pipelineLayouts[PIPELINE_TYPE_COUNT];
-    VkPipeline graphicsPipelines[PIPELINE_TYPE_COUNT];
+    std::array<VkPipelineLayout, PIPELINE_TYPE_COUNT> pipelineLayouts;
+    std::array<VkPipeline, PIPELINE_TYPE_COUNT> graphicsPipelines;
     VkDescriptorSetLayout descriptorSetLayout;
-    VkDescriptorSet descriptorSets[MAX_FRAMES_IN_FLIGHT];
+    std::array<VkDescriptorSet, MAX_FRAMES_IN_FLIGHT> descriptorSets;
     VkDescriptorPool descriptorPool;
     VkDevice device;
 
-    static std::vector<char> readFile(const std::string& filename) {
+    inline static std::vector<char> readFile(const std::string& filename) {
         std::string base = osd::getExecutableDir();
         std::ifstream file(std::filesystem::path(base) / filename, std::ios::ate | std::ios::binary);
